refactor(dashboard): request boards
