+++ src/modules/dashboard/Board/components/CardBoard/components/CardBoardAvatar/types/CardBoardAvatar.types.ts
@@ -1,4 +1,4 @@
-import { IMemberBoard } from "modules/dashboard/Board/components/CardBoard/services/types";
+import { IMemberBoard } from "modules/dashboard/Board/services/types";
 
 export interface ICardBoardAvatar {
 	membersBoard: Array<IMemberBoard>;
